move placeholder into atoms
