+++ react/src/components/atoms/Placeholder/index.js
