+++ react/src/components/templates/NarrowTemplate/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import SiteLogo from '../../atoms/media/SiteLogo';
-import Placeholder from '../../base/Placeholder';
+import Placeholder from '../../atoms/Placeholder';
 
 const NarrowTemplate = (narrowTemplate) => {
   const sideClass = narrowTemplate.side ? ` ma__narrow-template--${narrowTemplate.side}` : '';
